# expression: implement vectorized evaluation for builtinCeilDecToDecSi…