--- a/expression/builtin_math_vec.go
+++ b/expression/builtin_math_vec.go
@@ -865,24 +865,21 @@ func (b *builtinFloorDecToDecSig) vecEvalDecimal(input *chunk.Chunk, result *chu
 	if err := b.args[0].VecEvalDecimal(b.ctx, input, result); err != nil {
 		return err
 	}
-	result.ResizeDecimal(n, false)
-	result.MergeNulls(result)
 	res := result.Decimals()
-	bufs := result.Decimals()
 
 	for i := 0; i < n; i++ {
 		if result.IsNull(i) {
 			continue
 		}
 		rst := new(types.MyDecimal)
-		if bufs[i].IsNegative() {
+		if !res[i].IsNegative() {
 			err := res[i].Round(rst, 0, types.ModeTruncate)
 			if err != nil {
 				return err
 			}
 		}
 		err := res[i].Round(rst, 0, types.ModeTruncate)
-		if err != nil || rst.Compare(&bufs[i]) == 0 {
+		if err != nil || rst.Compare(&res[i]) == 0 {
 			if err != nil {
 				return err
 			}