@Zhuchengyu04 2019-11-13
> #13314 has been dealing with `builtinCeilDecToDecSig`, could you please fix the `builtinFloorDecToDecSig` in this PR?

With pleasure